+++ src/wsdd.py
@@ -287,8 +287,10 @@ class MulticastHandler:
         self.uc_send_socket.bind((self.address.address_str, WSD_UDP_PORT))
 
         self.mc_send_socket.setsockopt(socket.IPPROTO_IP, socket.IP_MULTICAST_IF, mreq)
-        self.mc_send_socket.setsockopt(socket.IPPROTO_IP, socket.IP_MULTICAST_LOOP, 0)
-        self.mc_send_socket.setsockopt(socket.IPPROTO_IP, socket.IP_MULTICAST_TTL, args.hoplimit)
+        # OpenBSD requires the optlen to be sizeof(char) for LOOP and TTL options
+        # (see also https://github.com/python/cpython/issues/67316)
+        self.mc_send_socket.setsockopt(socket.IPPROTO_IP, socket.IP_MULTICAST_LOOP, struct.pack('B', 0))
+        self.mc_send_socket.setsockopt(socket.IPPROTO_IP, socket.IP_MULTICAST_TTL, struct.pack('B', args.hoplimit))
 
         self.listen_address = (self.address.address_str, WSD_HTTP_PORT)
 
@@ -1580,7 +1582,7 @@ NET_RT_IFLIST: int = 3
 
 # from sys/net/if.h
 IFF_LOOPBACK: int = 0x8
-IFF_MULTICAST: int = 0x800
+IFF_MULTICAST: int = 0x800 if platform.system() != 'OpenBSD' else 0x8000
 
 # sys/netinet6/in6_var.h
 IN6_IFF_TENTATIVE: int = 0x02
@@ -1595,28 +1597,38 @@ class RouteSocketAddressMonitor(NetworkAddressMonitor):
     Implementation of the AddressMonitor for FreeBSD and Darwin using route sockets
     """
 
-    # Common definition for beginning part of if(m?a)?_msghdr structs (see net/if.h).
-    IF_COMMON_HDR_DEF: str = '@HBBii'
+    # Common definition for beginning part of if(m?a)?_msghdr structs (see net/if.h/man 4 route).
+    IF_COMMON_HDR_DEF = '@HBBii' if platform.system() != 'OpenBSD' else'@HBBHHHBBiii'
+
+    # from net/if.h
     RTM_NEWADDR: int = 0xC
     RTM_DELADDR: int = 0xD
+    # not tested for OpenBSD
     RTM_IFINFO: int = 0xE
 
+    # from route.h (value equals for FreeBSD, Darwin and OpenBSD)
+    RTM_VERSION: int = 0x5
+
     # from net/if.h (struct ifa_msghdr)
-    IFA_MSGHDR_DEF: str = IF_COMMON_HDR_DEF + 'hi'
+    IFA_MSGHDR_DEF: str = IF_COMMON_HDR_DEF + ('hi' if platform.system() != 'OpenBSD' else '')
     IFA_MSGHDR_SIZE: int = struct.calcsize(IFA_MSGHDR_DEF)
 
-    IF_MSGHDR_DEF_BASE: str = IF_COMMON_HDR_DEF + 'h'
     # The struct package does not allow to specify those, thus we hard code them as chars (x4).
-    IF_DATA_DEFS: Dict[str, str] = {
+    IF_MSG_DEFS: Dict[str, str] = {
         # if_data in if_msghdr is prepended with an u_short _ifm_spare1, thus the 'H' a the beginning)
-        'FreeBSD': 'H6c2c8c8c104c8c16c',
+        'FreeBSD': 'hH6c2c8c8c104c8c16c',
         # There are 8 bytes and 22 uint32_t in the if_data struct (22 x 4 Bytes + 8 = 96 Bytes)
         # It is also aligned on 4-byte boundary necessitating 2 bytes padding inside if_msghdr
-        'Darwin': '2c8c22I'
+        'Darwin': 'h2c8c22I',
+        # struct if_data from /src/sys/net/if.h for if_msghdr
+        #  (includes struct timeval which is a int64 + long
+        'OpenBSD': '4c3I13Q1Iql'
     }
 
     socket: socket.socket
     intf_blacklist: List[str]
+
+    is_openbsd: bool = False
 
     def __init__(self, aio_loop: asyncio.AbstractEventLoop) -> None:
         super().__init__(aio_loop)
@@ -1627,7 +1639,8 @@ class RouteSocketAddressMonitor(NetworkAddressMonitor):
         self.socket = socket.socket(socket.AF_ROUTE, socket.SOCK_RAW, socket.AF_UNSPEC)
         self.aio_loop.add_reader(self.socket.fileno(), self.handle_change)
 
-        self.IF_MSGHDR_SIZE = struct.calcsize(self.IF_MSGHDR_DEF_BASE + self.IF_DATA_DEFS[platform.system()])
+        self.IF_MSGHDR_SIZE = struct.calcsize(self.IF_COMMON_HDR_DEF + self.IF_MSG_DEFS[platform.system()])
+        self.is_openbsd = platform.system() == 'OpenBSD'
 
     def do_enumerate(self) -> None:
         super().do_enumerate()
@@ -1661,9 +1674,20 @@ class RouteSocketAddressMonitor(NetworkAddressMonitor):
         intf_flags = 0
         while offset < len(buf):
             # unpack route message response
-            rtm_len, _, rtm_type, addr_mask, flags = struct.unpack_from(self.IF_COMMON_HDR_DEF, buf, offset)
+            if not self.is_openbsd:
+                rtm_len, rtm_version, rtm_type, addr_mask, flags = struct.unpack_from(
+                        self.IF_COMMON_HDR_DEF, buf, offset)
+            else:
+                rtm_len, rtm_version, rtm_type, ifa_hdr_len, _, _, _, _, addr_mask, flags, _ = struct.unpack_from(
+                        self.IF_COMMON_HDR_DEF, buf, offset)
 
-            if rtm_type not in [self.RTM_NEWADDR, self.RTM_DELADDR, self.RTM_IFINFO]:
+            # exit condition for OpenBSD where always the complete buffer (ie 4096 bytes) is returned
+            if rtm_len == 0:
+                break;
+
+            # skip over non-understood packets and versions
+            if (rtm_type not in [self.RTM_NEWADDR, self.RTM_DELADDR, self.RTM_IFINFO]) or (
+                    rtm_version != self.RTM_VERSION):
                 offset += rtm_len
                 continue
 
@@ -1946,7 +1970,7 @@ def drop_privileges(uid: int, gid: int) -> bool:
 def create_address_monitor(system: str, aio_loop: asyncio.AbstractEventLoop) -> NetworkAddressMonitor:
     if system == 'Linux':
         return NetlinkAddressMonitor(aio_loop)
-    elif system in ['FreeBSD', 'Darwin']:
+    elif system in ['FreeBSD', 'Darwin', 'OpenBSD']:
         return RouteSocketAddressMonitor(aio_loop)
     else:
         raise NotImplementedError('unsupported OS: ' + system)