+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Support for OpenBSD (tested on riscv64 with OpenBSD 7.4)
 - Configuration files for firewalld (#186). Thanks to Ondrej Holy.
 - Show device type and allow filtering in API's `list` command (#189). Thanks to Ondrej Holy.
 - Add option `--metadata-timeout` to set the timeout for the HTTP-based metadata exchange (closes #83)